+++ src/helper/object.js
@@ -103,16 +103,21 @@ function toSnakeCase(object, exceptions) {
   }, {});
 }
 
-function toCamelCase(object, exceptions) {
+function toCamelCase(object, exceptions, options) {
   if (typeof object !== 'object' || assert.isArray(object) || object === null) {
     return object;
   }
 
   exceptions = exceptions || [];
-
+  options = options || {};
   return Object.keys(object).reduce(function(p, key) {
     var newKey = exceptions.indexOf(key) === -1 ? snakeToCamel(key) : key;
-    p[newKey] = toCamelCase(object[key]);
+
+    p[newKey] = toCamelCase(object[newKey] || object[key], [], options);
+
+    if (options.keepOriginal) {
+      p[key] = toCamelCase(object[key], [], options);
+    }
     return p;
   }, {});
 }
@@ -147,6 +152,22 @@ function getOriginFromUrl(url) {
   return origin;
 }
 
+function trim(options, key) {
+  var trimmed = extend(options);
+  if (options[key]) {
+    trimmed[key] = options[key].trim();
+  }
+  return trimmed;
+}
+
+function trimMultiple(options, keys) {
+  return keys.reduce(trim, options);
+}
+
+function trimUserDetails(options) {
+  return trimMultiple(options, ['username', 'email', 'phoneNumber']);
+}
+
 export default {
   toSnakeCase: toSnakeCase,
   toCamelCase: toCamelCase,
@@ -156,5 +177,6 @@ export default {
   getKeysNotIn: getKeysNotIn,
   extend: extend,
   getOriginFromUrl: getOriginFromUrl,
-  getLocationFromUrl: getLocationFromUrl
+  getLocationFromUrl: getLocationFromUrl,
+  trimUserDetails: trimUserDetails
 };