+++ src/web-auth/hosted-pages.js
@@ -13,6 +13,7 @@ import paramsArray from '../helper/param-constants';
 function HostedPages(client, options) {
   this.baseOptions = options;
   this.client = client;
+  this.baseOptions.universalLoginPage = true;
   this.request = new RequestBuilder(this.baseOptions);
 
   this.warn = new Warn({