+++ src/web-auth/popup.js
@@ -177,7 +177,12 @@ Popup.prototype.authorize = function(options, cb) {
   }
 
   if (options.popupOptions) {
-    popOpts.popupOptions = objectHelper.pick(options.popupOptions, ['width', 'height']);
+    popOpts.popupOptions = objectHelper.pick(options.popupOptions, [
+      'width',
+      'height',
+      'top',
+      'left'
+    ]);
   }
 
   if (pluginHandler) {
@@ -192,7 +197,7 @@ Popup.prototype.authorize = function(options, cb) {
 
   popup = this.getPopupHandler(options);
 
-  return popup.load(url, relayUrl, popOpts, responseHandler(cb));
+  return popup.load(url, relayUrl, popOpts, responseHandler(cb, { keepOriginalCasing: true }));
 };
 
 /**